Add a first draft of the activity service
This service is a way of counting visitors in real-time, and intended to
replace the Cassandra-based storage of active visitors in subreddits and
live threads.

This service is made of two components: Thrift and HTTP. The thrift
service is entirely internal, while the HTTP one can be hit by users
directly. We won't use the HTTP portion immediately, but it is intended
to allow logged-out users to be counted as well (including for reddit
live).