Always fuzz activity counts, even at higher values
Jesse Ruderman reported that by maintaining a stable number of bots in a
given subreddit, the fuzzing can be subverted to determine when a
specific targeted user is directed to that subreddit. We could boost the
fuzzing threshold to a less feasibly mimiced count, but there's little
value in the number being super precise.